moved _apply_operation and _handle_finite_sets
