+++ sympy/sets/setexpr.py
@@ -7,9 +7,6 @@ from sympy.utilities.iterables import sift
 from sympy.multipledispatch import dispatch, Dispatcher
 from sympy.sets.sets import set_add, set_sub, set_mul, set_div, set_pow, set_function
 
-d = Dummy('d')
-
-
 class SetExpr(Expr):
     """An expression that can take on values of a set
 
@@ -101,32 +98,3 @@ def _setexpr_apply_operation(op, x, y):
         y = y.set
     out = op(x, y)
     return SetExpr(out)
-
-
-def _handle_finite_sets(op, x, y, commutative):
-    # Handle finite sets:
-    fs_args, other = sift([x, y], lambda x: isinstance(x, FiniteSet), binary=True)
-    if len(fs_args) == 2:
-        return FiniteSet(*[op(i, j) for i in fs_args[0] for j in fs_args[1]])
-    elif len(fs_args) == 1:
-        sets = [_apply_operation(op, other[0], i, commutative) for i in fs_args[0]]
-        return Union(*sets)
-    else:
-        return None
-
-def _apply_operation(op, x, y, commutative):
-    out = _handle_finite_sets(op, x, y, commutative)
-    if out is None:
-        out = op(x, y)
-
-    if out is None and commutative:
-        out = op(y, x)
-    if out is None:
-        _x, _y = symbols("x y")
-        if isinstance(x, Set) and not isinstance(y, Set):
-            out = ImageSet(Lambda(d, op(d, y)), x).doit()
-        elif not isinstance(x, Set) and isinstance(y, Set):
-            out = ImageSet(Lambda(d, op(x, d)), y).doit()
-        else:
-            out = ImageSet(Lambda((_x, _y), op(_x, _y)), x, y)
-    return out
